Example of moving component operations to peregrine
