+++ packages/venia-ui/lib/components/CommunicationsPage/__tests__/communicationsPage.spec.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { createTestInstance } from '@magento/peregrine';
-import { useCommunicationsPage } from '@magento/peregrine/lib/talons/MyAccount/useCommunicationsPage';
+import { useCommunicationsPage } from '@magento/peregrine/lib/talons/CommunicationsPage/useCommunicationsPage';
 
 import CommunicationsPage from '../communicationsPage';
 import LoadingIndicator from '../../LoadingIndicator';
